add class methods to oad and save lookup table
A pickle file will be dumped to memory for ease of use with lookup tables,
e.g. with converting latlon coordinates if the sensor configurations haven't
changed. Currently there is no way to change the path to the lookup data.

The save function first checks if there is any change in the lookup data,
and if there is, only then is it written.